Bug 611 - serialization of path attributes - added Nlri,originatorId and clusterId serialization to BGPUPdateMessageParser - added Ipv4NextHopCase Ipv6NextHopCase serialization - added serialization methods to attribute objects - attribute_flags and attribute_type added to all attribute serializers - added tests for Update message serialization

+++ bgp/parser-impl/src/main/java/org/opendaylight/protocol/bgp/parser/impl/message/update/AsPathSegmentParser.java
@@ -91,7 +91,7 @@ public final class AsPathSegmentParser {
         byteAggregator.writeByte((byte) aSetCase.getASet().getAsSet().size());
         for (AsNumber asNumber : aSetCase.getASet().getAsSet()) {
             ShortAsNumber shortAsNumber = new ShortAsNumber(asNumber);
-            byteAggregator.writeShort(shortAsNumber.getValue().shortValue());
+            byteAggregator.writeInt(shortAsNumber.getValue().intValue());
         }
     }
 
@@ -103,7 +103,7 @@ public final class AsPathSegmentParser {
         byteAggregator.writeByte((byte) aListCase.getAList().getAsSequence().size());
         for (AsSequence value : aListCase.getAList().getAsSequence()) {
             ShortAsNumber shortAsNumber = new ShortAsNumber(value.getAs());
-            byteAggregator.writeShort(shortAsNumber.getValue().shortValue());
+            byteAggregator.writeInt(shortAsNumber.getValue().intValue());
         }
     }
 }
